Separate Mesh/Submesh into separate files
